Fix web3 helpers
