+++ test/helpers/web3-helpers.js
@@ -1,0 +1,24 @@
+const getSignatureFields = signature => {
+  signature = signature.slice(2) // remove 0x
+  const v = toDecimal(`0x${signature.slice(128, 130)}`)
+
+  return {
+    v: v != 27 && v != 28 ? v + 27 : v,
+    r: `0x${signature.slice(0, 64)}`,
+    s: `0x${signature.slice(64, 128)}`,
+  }
+}
+
+// in versions below 1.0.0 utils not defined
+const sign = (...args) => (web3.utils ? web3.eth.sign(...args) : web3.eth.sign(...args.reverse()))
+
+const toDecimal = web3.utils ? web3.utils.toDecimal : web3.toDecimal
+
+const sha3 = web3.utils ? web3.utils.sha3 : web3.sha3
+
+module.exports = {
+  getSignatureFields,
+  sha3,
+  toDecimal,
+  sign,
+}
